Move logger testing fixture to conftest since it's usable multiple places
