+++ api/conftest.py
@@ -1,3 +1,4 @@
+import logging
 import os
 import typing
 from unittest.mock import MagicMock
@@ -1072,3 +1073,20 @@ def superuser():
 def superuser_client(superuser: FFAdminUser, client: APIClient):
     client.force_login(superuser, backend="django.contrib.auth.backends.ModelBackend")
     return client
+
+
+@pytest.fixture
+def inspecting_handler() -> logging.Handler:
+    """
+    Fixture used to test the output of logger related output.
+    """
+
+    class InspectingHandler(logging.Handler):
+        def __init__(self, *args, **kwargs) -> None:
+            super().__init__(*args, **kwargs)
+            self.messages = []
+
+        def handle(self, record: logging.LogRecord) -> None:
+            self.messages.append(self.format(record))
+
+    return InspectingHandler()
